requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cryptography==3.4.8
cuda-python==12.5.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distro==1.7.0
distro-info==1.0
dm-reverb==0.14.0
dm-tree==0.1.8
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
flatbuffers==24.3.25
fonttools==4.53.0
gast==0.5.4
gin-config==0.5.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
gym==0.23.0
gym-notices==0.0.8
h5py==3.11.0
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==2.15.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
ml-dtypes==0.2.0
more-itertools==8.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.4
opt-einsum==3.3.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
portpicker==1.6.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pygame==2.1.3
Pygments==2.18.0
PyJWT==2.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rlds==0.1.8
rpds-py==0.19.0
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-probability==0.23.0
termcolor==2.4.0
tf-agents==0.19.0
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.5.0
tzdata==2024.1
urllib3==2.2.2
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.14.1
yarg==0.1.9
zipp==1.0.0
